+++ packages/lib/src/types/objectBased/typesArraySet.ts
@@ -47,7 +47,7 @@ export function typesArraySet<T extends AnyType>(valueType: T): ModelType<ArrayS
     const thisTc: TypeChecker = new TypeChecker(
       TypeCheckerBaseType.Object, // because it is really a model
 
-      (obj, path, typeCheckedValue) => {
+      (obj, path, typeCheckedValue, partialCheckScope) => {
         if (!(obj instanceof ArraySet)) {
           return new TypeCheckError({
             path,
@@ -58,8 +58,9 @@ export function typesArraySet<T extends AnyType>(valueType: T): ModelType<ArrayS
         }
 
         const resolvedTc = resolveTypeChecker(dataTypeChecker)
-        return resolvedTc.check(obj.$, path, typeCheckedValue)
+        return resolvedTc.check(obj.$, path, typeCheckedValue, partialCheckScope)
       },
+      undefined,
 
       getTypeName,
       typeInfoGen,